% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_actions.R
\name{add_CV_item}
\alias{add_CV_item}
\alias{add_CV_list}
\title{Lower level functions to enter activities in database}
\usage{
add_CV_item(table, key, ...)

add_CV_list(table, key, ...)
}
\arguments{
\item{table}{Name of table.}

\item{key}{Name of the skill.}

\item{...}{Items to be included.}
}
\value{
Invisibly returns NULL.
}
\description{
Lower level functions to enter activities in database
}
\examples{
if (interactive()) {
 init()
 # create new table
 DBI::dbCreateTable(get_connection(), "hobbies", c(skill = "TEXT", item = "TEXT"))
 # add a list of activities
 add_CV_list("hobbies", key = NULL, sport, c("tennis", "rugby"))
}
}
