% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_topics.R
\name{add_education}
\alias{add_education}
\alias{add_experience}
\alias{add_skills}
\title{Predefined functions for database entry}
\usage{
add_education(begin, end, place, title, type, degree, key = NULL)

add_experience(begin, end, position, place, description, key = NULL)

add_skills(skill, ...)
}
\arguments{
\item{begin}{Begin date of entry.}

\item{end}{End date of entry.}

\item{place}{Place of activity.}

\item{title}{Title of activity.}

\item{type}{Type of activity.}

\item{degree}{Degree in case of education.}

\item{key}{The key can be used to prevent double entries and takes a string
or a vector of strings that represent the variable to be used as unique
identifier.}

\item{position}{Title of the position.}

\item{description}{Description of the position.}

\item{skill}{Name of the skill.}

\item{...}{In case of \code{add_skills()} takes a list of skills.}
}
\value{
Invisibly returns NULL.
}
\description{
These functions are meant to exemplify the functionality of \code{staticv}. More
flexibility can be gained by using the lower level functions \code{add_CV_item()}
for ordered (possibly chronological) itemized entries and \code{add_CV_list()} for
un-ordered lists of information.
}
\examples{
if (interactive()) {
init()
add_education(2010, 2014, "University of Musterhaussen", "Study of Information Technology", "degree", "Master of Science", key = c("begin", "end"))
add_education(2006, 2010, "University of Musterhaussen", "Study of Social Sciences", "degree", "Bachelor of Science", key = c("begin", "end"))
add_education(2000, 2006, "Musterstadt Public Schhool", "Gymnasium", "degree", "Diploma", key = c("begin", "end"))
}

}
