% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{get_connection}
\alias{get_connection}
\title{Get database connection}
\usage{
get_connection()
}
\value{
Returns database connection.
}
\description{
Get database connection
}
\examples{
if (interactive()) {
 get_connection()
}
}
