% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{get_CV_table}
\alias{get_CV_table}
\alias{delete_CV_table}
\title{Retrieves/delete tables from database.}
\usage{
get_CV_table(table)

delete_CV_table(table, verbose = TRUE)
}
\arguments{
\item{table}{Name of table.}

\item{verbose}{Set to \code{FALSE} to make function silent.}
}
\value{
Invisibly returns NULL.
}
\description{
Retrieves/delete tables from database.
}
\examples{
if (interactive()) {
 init()
 get_CV_table("education")
}
}
