% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{init}
\alias{init}
\title{Initialize database.}
\usage{
init()
}
\arguments{
\item{name}{Short name for sqlite database.}
}
\value{
Invisibly returns NULL.
}
\description{
Initialize database.
}
\examples{
if (interactive()) {
 init()
}
}
