% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_styling.R
\name{reformat_cv_table}
\alias{reformat_cv_table}
\title{Convert datatables from database to long format.}
\usage{
reformat_cv_table(topic, key, body, sep = "--")
}
\arguments{
\item{topic}{\code{data.frame()} to be converted to long format.}

\item{key}{Character string or vector of variables included in key. In case
of multiple variables these are required to be of same type.}

\item{body}{Character string or vector of variables included in body of
table. In case of multiple variables these are required to be of same type.}

\item{sep}{Separator in case of multiple variable used as key.}
}
\value{
\code{data.frame()} to be prepared for inclusion in the CV.
}
\description{
Convert datatables from database to long format.
}
\examples{
reformat_cv_table(iris, "Sepal.Length", c("Sepal.Width", "Petal.Width"))
}
