% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticv-document.R
\name{staticv}
\alias{staticv}
\title{Staticv document}
\usage{
staticv(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Other parameters to generate PDF document with rmarkdown render.}

\item{keep_tex}{Keep text file.}
}
\description{
Staticv document
}
